nbformat downgrade 5.3.0
This is a temporary downgrade to immediately fix the rendering warning/error breaking the export cell on our notebooks. The problem is that one or more of the notebook cells do not have the mandatory "id" field. I will be updating the notebooks themselves next week and we will figure out a plan from there.